+++ OptimoveCore/Sources/Classes/Constants/SDKVersion.swift
@@ -1,3 +1,2 @@
 //  Copyright © 2019 Optimove. All rights reserved.
-
-public let SDKVersion = "6.9.0"
+public let SDKVersion = "6.11.0"